+++ source/Modules/index.js
@@ -1,4 +1,1 @@
 export * from './storageManager';
-export * from './Controller/connection';
-export * from './Controller/transaction';
-export * from './Controller/information';
