+++ source/views/Page/index.jsx
@@ -6,7 +6,7 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 import { theme } from '@components';
-import initConfig from '../locales';
+import initConfig from '../../locales';
 import Welcome from './Views/Welcome';
 
 i18n.use(initReactI18next).init(initConfig);
