+++ source/views/ProviderPopups/Notification/components/Transfer/index.jsx
@@ -9,7 +9,7 @@ import PropTypes from 'prop-types';
 import { useDispatch } from 'react-redux';
 import { HANDLER_TYPES, sendMessage } from '@background/Keyring';
 import { setAccountInfo } from '@redux/wallet';
-import initConfig from '../../../../locales';
+import initConfig from '../../../../../locales';
 import ErrorScreen from '../NotificationError';
 import useStyles from './styles';
 import RequestHandler from './components/RequestHandler';
